+++ wallet/res/menu/payment_options.xml
@@ -5,7 +5,7 @@
     <item
         android:id="@+id/option_close"
         android:icon="@drawable/ic_close_blue"
-        android:title="Close"
+        android:title="@string/payment_request_close"
         android:visible="true"
         app:showAsAction="always" />
